Streamline codebase: fix reproducibility bug, type hints, deduplication, tests, CI
Fix MatrixFeedback.duel_repeatedly using global np.random instead of
self.random_state, breaking reproducibility with fixed seeds. Replace
deprecated np.float/np.bool aliases. Standardize Optional[RandomState]
across 9 algorithm constructors (several had mutable RandomState()
defaults). Replace np.array type annotations with np.ndarray throughout.
Deduplicate get_winners_against/get_losers_against via shared helper.
Replace assert with ValueError in Regret.__init__. Add 48 unit tests
across 6 new files covering core stats, feedback, and utility classes.
Fix CI Sphinx duplication, add Python 3.8 test job, coverage reporting,
and replace deprecated setup.py build with python -m build. Expand README
with installation, quick-start example, and dev setup instructions.

+++ duelpy/algorithms/rmed.py
@@ -123,7 +123,7 @@ class Rmed1(Algorithm):
         self,
         feedback_mechanism: FeedbackMechanism,
         time_horizon: int,
-        divergence_tolerance: Optional[Callable[[int], np.float]] = None,
+        divergence_tolerance: Optional[Callable[[int], float]] = None,
         random_state: Optional[np.random.RandomState] = None,
     ):
         super().__init__(
@@ -431,7 +431,7 @@ class Rmed2(Rmed1):
         self,
         feedback_mechanism: FeedbackMechanism,
         time_horizon: int,
-        divergence_tolerance: Optional[Callable[[int], np.float]] = None,
+        divergence_tolerance: Optional[Callable[[int], float]] = None,
         random_state: Optional[np.random.RandomState] = None,
         exploratory_constant: float = 3,
     ):
@@ -632,7 +632,7 @@ class Rmed2FH(Rmed2):
         self,
         feedback_mechanism: FeedbackMechanism,
         time_horizon: int,
-        divergence_tolerance: Optional[Callable[[int], np.float]] = None,
+        divergence_tolerance: Optional[Callable[[int], float]] = None,
         random_state: Optional[np.random.RandomState] = None,
         exploratory_constant: float = 3,
     ):
